+++ .gitignore
@@ -19,6 +19,7 @@ wheels/
 *.egg-info/
 .installed.cfg
 *.egg
+_delete_/
 
 # Virtual Environment
 .env
